refactor: Rename utility paths and update state management in DeveloperDock and dataSlice
