+++ src/features/dashboard/utils/cn.ts
@@ -1,3 +1,3 @@
-export default function cn(...classes) {
+export default function cn(...classes: any) {
   return classes.filter(Boolean).join(' ');
 }
