+++ src/features/dev-dock/utils/data-slice.ts
@@ -2,15 +2,15 @@ import { createSlice } from '@reduxjs/toolkit';
 import { ViewMode } from '../data/type';
 
 const initialState = {
-    dataViewMode: ViewMode.FAKE_DATA,
+    current: ViewMode.FAKE_DATA,
 };
 
 export const dataSlice = createSlice({
-    name: 'data',
+    name: 'dataViewMode',
     initialState,
     reducers: {
         setDataViewMode: (state, action) => {
-            state.dataViewMode = action.payload;
+            state.current = action.payload;
         },
     },
 });
